+++ src/main/java/com/leonlib/repository/BookImagesRepository.java
@@ -8,5 +8,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import com.leonlib.model.BookImageInfo;
 
 public interface BookImagesRepository extends JpaRepository<BookImageInfo, Long> {
-    Optional<List<BookImageInfo>> findByBookId(Long bookId);
+    Optional<List<BookImageInfo>> findByBookId(Integer bookId);
 }
